% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_colours.R
\name{hj_colours}
\alias{hj_colours}
\title{HJ Colours}
\usage{
hj_colours(...)
}
\arguments{
\item{...}{Names of the colours}
}
\description{
Get hex codes for plot colours.
}
\details{
Use \code{hj_colours()} to see the full list of colours. This list
comes from
\href{https://registry.origami.ft.com/components/o-colors}{Origami
o-colors}.
}
\examples{
# Full list
hj_colours()

# Choose colours
hj_colours("paper")
hj_colours("oxford", "claret")

}
