% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_colours.R
\name{hj_cols}
\alias{hj_cols}
\title{HJ Colours Function to Select Colour from Palette}
\usage{
hj_cols(...)
}
\arguments{
\item{...}{Names of the colours}
}
\description{
Get hex codes for plot colours.
}
\details{
Use \code{hj_cols()} to see the full list of colours.
}
\examples{
# Full list
hj_cols()

# Choose colours
hj_cols("paper")
hj_cols("oxford", "claret")

}
