% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_colours.R
\name{hj_pal}
\alias{hj_pal}
\title{HJ Colour Palettes Generator.}
\usage{
hj_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Returns a function that can interplotate between colors in a palette.
}
