% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hj_theme.R
\name{hj_theme}
\alias{hj_theme}
\title{ggplot2 theme for simple, functional and beautiful graphs.}
\usage{
hj_theme(
  legend_right = FALSE,
  base_size = 12,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{legend_right}{Logical indicating whether legend should be placed to
the right of the plot. If FALSE, the default, legend is positioned above the
plot. The caption will change position relative to where the legend is
positioned.}

\item{base_size}{The base font size}

\item{base_family}{Font family}

\item{base_line_size}{Default}

\item{base_rect_size}{Default}
}
\description{
ggplot2 theme for simple, functional and beautiful graphs.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  hj_theme()

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  facet_wrap(vars(class)) +
  hj_theme()

}
